feat: 사이드바 설정 UI
